Fix player_wizardry
The check for player species being Astral was causing this to always
return 0 for non-Astrals. Explicitly cast this subexpression to int.

Fixes #52.

+++ crawl-ref/source/player.cc
@@ -2442,7 +2442,7 @@ int player_wizardry(spell_type /*spell*/)
 {
     return you.wearing(EQ_RINGS, RING_WIZARDRY)
            + you.wearing(EQ_STAFF, STAFF_WIZARDRY)
-           + you.species == SP_ASTRAL;
+           + (int)(you.species == SP_ASTRAL);
 }
 
 /**
